chore: move Release, Updater, UpdaterInfo, ReleaseTagInfo and SemanticVersion to YAT.Update namespace
